requirements.txt
aiogram==3.20.0.post0
tortoise-orm==0.21.7
pydantic==2.7.4
humanize==4.12.3
environs==11.0.0